% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_convolution_density.R
\name{Beta_convolution_density}
\alias{Beta_convolution_density}
\title{Calculate 4 parameter Beta PDF}
\usage{
Beta_convolution_density(z, alpha_matrix, beta_matrix, weighted_samps, weights)
}
\arguments{
\item{z}{evaluation points}

\item{alpha_matrix}{matrix of shape parameters for each element (b) in the aggregate over each of the N observations (N rows x b columns)}

\item{beta_matrix}{matrix of shape parameters for each element in the aggregate (N rows x b columns)}

\item{weighted_samps}{matrix of weighted samples for each element in the aggregate (N rows x b columns)}

\item{weights}{vector of weights that are used to combine to form the aggregated density Z. (length b)}
}
\value{
The aggregate density Z over a grid of values.
}
\description{
Calculate 4 parameter Beta PDF
}
\details{
A wrapper function for calculating Monte Carlo estimates for the aggregated Beta density Z using \code{ExtDist::dBeta_ab}.
}
