% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_convolution_density_parallel.R
\name{Beta_convolution_density_parallel}
\alias{Beta_convolution_density_parallel}
\title{Parallelized normalised predictive density over vector z}
\usage{
Beta_convolution_density_parallel(
  z_values,
  alpha_matrix,
  beta_matrix,
  weighted_samps,
  weights
)
}
\arguments{
\item{z_values}{evaluation points}

\item{alpha_matrix}{matrix of shape parameters for each element (b) in the aggregate over each of the N observations (N rows x b columns)}

\item{beta_matrix}{matrix of shape parameters for each element in the aggregate (N rows x b columns)}

\item{weighted_samps}{matrix of weighted samples for each element in the aggregate (N rows x b columns)}

\item{weights}{vector of weights that are used to combine to form the aggregated density Z. (length b)}
}
\value{
The aggregate density Z over a grid of values.
}
\description{
Parallelized normalised predictive density over vector z
}
\details{
A wrapper function for calculating Monte Carlo estimates for the aggregate density Z.
}
