% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_sim_data.R
\docType{data}
\name{beta_sim_data}
\alias{beta_sim_data}
\title{Simulated bounded hierarchical time series data}
\format{
A data.frame with columns:
\describe{
\item{Time}{Time index}
\item{A}{Top-level proportion}
\item{AA}{Bottom-level proportion (node AA)}
\item{AB}{Bottom-level proportion (node AB)}
}
}
\source{
Simulated by the package authors
}
\usage{
beta_sim_data
}
\description{
A simulated dataset used to illustrate hierarchical coherent
forecasting for bounded time series.
}
\keyword{datasets}
