% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBeta_4p.R
\name{dBeta_4p}
\alias{dBeta_4p}
\title{Calculate 4 parameter Beta PDF}
\usage{
dBeta_4p(x, lower = 0, upper, alpha, beta)
}
\arguments{
\item{x}{evaluation points}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{alpha}{estimated shape parameter of Beta distribution}

\item{beta}{estimated shape parameter of Beta distribution}
}
\value{
The density of the 4 parameter Beta distribution
}
\description{
Calculate 4 parameter Beta PDF
}
\details{
A wrapper function for \code{ExtDist::dBeta_ab}.
}
\examples{
\dontrun{
dBeta_4p(x = 0.07, lower=0, upper=0.5, alpha=5, beta=2)
}
}
