% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZIB_4p.R
\name{dZIB_4p}
\alias{dZIB_4p}
\title{Vectorized Zero Inflated 4 parameter Beta density}
\usage{
dZIB_4p(z, Y_mc, phi_mc, zi_mc, upper, lower = 0)
}
\arguments{
\item{z}{evaluation points}

\item{Y_mc}{Monte Carlo draws of Y}

\item{phi_mc}{Monte Carlo draws of Beta precision (n_draws x n_nodes x n_years)}

\item{zi_mc}{Monte Carlo draws of zero inflation (n_draws x n_nodes x n_years)}

\item{upper}{upper bound}

\item{lower}{lower bound}
}
\description{
Vectorized Zero Inflated 4 parameter Beta density
}
