% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZOIB_4p.R
\name{dZOIB_4p}
\alias{dZOIB_4p}
\title{Vectorized ZOIB density}
\usage{
dZOIB_4p(z, Y_mc, phi_mc, zoi_mc, coi_mc, upper, lower = 0)
}
\arguments{
\item{z}{evaluation points}

\item{Y_mc}{Monte Carlo draws of Y}

\item{phi_mc}{Monte Carlo draws of phi}

\item{zoi_mc}{Monte Carlo draws of zero-one inflation}

\item{coi_mc}{Monte Carlo draws of conditional one inflation}

\item{upper}{upper bound}

\item{lower}{lower bound}
}
\description{
Vectorized ZOIB density
}
