% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_vcov_delta.R
\name{logit_vcov_delta}
\alias{logit_vcov_delta}
\title{Get logit transformation of Sigma variance-covariance matrix via Delta method}
\usage{
logit_vcov_delta(mu, Sigma)
}
\arguments{
\item{mu}{estimated mean}

\item{Sigma}{estimated variance-covariance matrix}
}
\value{
logit transformed variance-covariance matrix
}
\description{
Get logit transformation of Sigma variance-covariance matrix via Delta method
}
\details{
The logit transformation of Sigma variance-covariance matrix via the Delta method
}
