% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_condition_tilting.R
\name{moment_condition_tilting}
\alias{moment_condition_tilting}
\title{The moment condition via the cumulant generating function (CGF) for continuous densities.}
\usage{
moment_condition_tilting(nu, f_y, y_vals, mu_theory)
}
\arguments{
\item{nu}{The tilting parameter}

\item{f_y}{calculate density of aggregated}

\item{y_vals}{evaluation points over which the density is calculated}

\item{mu_theory}{The theorised (target) mean}
}
\value{
The residuals in means (estimated - target) after tilting with a given nu
}
\description{
The moment condition via the cumulant generating function (CGF) for continuous densities.
}
\details{
A loss function to optimise over to evaluate the tilting parameter (nu) used in exponential tilting.
}
