% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_sim_data.R
\docType{data}
\name{poisson_sim_data}
\alias{poisson_sim_data}
\title{Simulated bounded hierarchical time series data}
\format{
A data.frame with columns:
\describe{
\item{X}{Covariate}
\item{Tot}{Top-level count}
\item{AA}{Bottom-level count (node AA)}
\item{AB}{Bottom-level count (node AB)}
\item{BA}{Bottom-level count (node BA)}
\item{BB}{Bottom-level count (node BB)}
}
}
\source{
Simulated by the package authors
}
\usage{
poisson_sim_data
}
\description{
A simulated dataset used to illustrate hierarchical coherent
forecasting for bounded time series.
}
\keyword{datasets}
