% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBeta_4p.R
\name{rBeta_4p}
\alias{rBeta_4p}
\title{Vectorized 4 parameter Beta sampler using \code{ExtDist::rBeta_ab}.}
\usage{
rBeta_4p(n_mc, sub_obs_data, phi_array, weights)
}
\arguments{
\item{n_mc}{number of Monte Carlo samples}

\item{sub_obs_data}{matrix of bottom-level series to use as expectation in sampler (n_years x n_nodes)}

\item{phi_array}{array (n_draws x n_nodes x n_years)}

\item{weights}{node weights (vector of length n_nodes)}
}
\value{
3D array (n_draws x n_nodes x n_years)
}
\description{
Vectorized 4 parameter Beta sampler using \code{ExtDist::rBeta_ab}.
}
