% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZIB_4p.R
\name{rZIB_4p}
\alias{rZIB_4p}
\title{Vectorized Zero Inflated 4 parameter Beta sampler}
\usage{
rZIB_4p(n_mc, sub_obs_data, phi_array, zi_array, weights)
}
\arguments{
\item{n_mc}{number of Monte Carlo samples}

\item{sub_obs_data}{matrix of observed data (n_years x n_nodes)}

\item{phi_array}{Monte Carlo draws of Beta precision (n_draws x n_nodes x n_years)}

\item{zi_array}{Monte Carlo draws of zero inflation (n_draws x n_nodes x n_years)}

\item{weights}{node weights for 4 parameter Beta (vector of length n_nodes)}
}
\value{
3D array (n_draws x n_nodes x n_years)
}
\description{
Vectorized Zero Inflated 4 parameter Beta sampler
}
