% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZOIB_4p.R
\name{rZOIB_4p}
\alias{rZOIB_4p}
\title{Vectorized Zero-One Inflated 4 parameter Beta sampler}
\usage{
rZOIB_4p(n_mc, sub_obs_data, phi_array, zoi_array, coi_array, weights)
}
\arguments{
\item{n_mc}{number of Monte Carlo samples}

\item{sub_obs_data}{matrix of observed data (n_years x n_nodes)}

\item{phi_array}{array (n_draws x n_nodes x n_years)}

\item{zoi_array}{array (n_draws x n_nodes x n_years)}

\item{coi_array}{array (n_draws x n_nodes x n_years)}

\item{weights}{node weights (vector of length n_nodes)}
}
\value{
3D array (n_draws x n_nodes x n_years)
}
\description{
Vectorized Zero-One Inflated 4 parameter Beta sampler
}
