% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilted_density_discrete.R
\name{tilted_density_discrete}
\alias{tilted_density_discrete}
\title{Tilting the density of f(y) using the optimised nu parameter.}
\usage{
tilted_density_discrete(nu, f_y, y_vals)
}
\arguments{
\item{nu}{The tilting parameter}

\item{f_y}{calculate density of aggregated}

\item{y_vals}{evaluation points over which the density is calculated}
}
\value{
The tilted density of f(y).
}
\description{
Tilting the density of f(y) using the optimised nu parameter.
}
