% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIB_convolution_density.R
\name{ZIB_convolution_density}
\alias{ZIB_convolution_density}
\title{Vectorized Zero Inflated 4 parameter Beta density}
\usage{
ZIB_convolution_density(Y_mc, phi_array, zi_array, weights, z_values, n_mc)
}
\arguments{
\item{Y_mc}{Monte Carlo draws of weighted bottom-series samples}

\item{phi_array}{Monte Carlo draws of phi}

\item{zi_array}{Monte Carlo draws of zero inflation}

\item{weights}{node weights (vector of length n_nodes)}

\item{z_values}{evaluation points}

\item{n_mc}{number of Monte Carlo samples}
}
\description{
Vectorized Zero Inflated 4 parameter Beta density
}
