% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZOIB_convolution_density.R
\name{ZOIB_convolution_density}
\alias{ZOIB_convolution_density}
\title{Vectorized ZOIB density}
\usage{
ZOIB_convolution_density(
  Y_mc,
  phi_array,
  zoi_array,
  coi_array,
  weights,
  z_values,
  n_mc
)
}
\arguments{
\item{Y_mc}{Monte Carlo draws of weighted bottom-series samples}

\item{phi_array}{Monte Carlo draws of phi}

\item{zoi_array}{Monte Carlo draws of zero-one inflation}

\item{coi_array}{Monte Carlo draws of conditional one inflation}

\item{weights}{node weights (vector of length n_nodes)}

\item{z_values}{evaluation points}

\item{n_mc}{number of Monte Carlo samples}
}
\description{
Vectorized ZOIB density
}
